优化了代码逻辑结构，主函数更简洁
